[REFACTOR] Measure 제외한 Viewmodel 통합
