--- app/src/main/java/com/preonboarding/sensordashboard/presentation/sensor_history_play/SensorHistoryPlayViewModel.kt
@@ -1,7 +1,0 @@
-package com.preonboarding.sensordashboard.presentation.sensor_history_play
-
-import com.preonboarding.sensordashboard.common.base.BaseViewModel
-import javax.inject.Inject
-
-class SensorHistoryPlayViewModel : BaseViewModel() {
-}
