--- app/src/main/java/com/preonboarding/sensordashboard/presentation/sensor_history_show/SensorHistoryShowViewModel.kt
@@ -1,6 +1,0 @@
-package com.preonboarding.sensordashboard.presentation.sensor_history_show
-
-import com.preonboarding.sensordashboard.common.base.BaseViewModel
-
-class SensorHistoryShowViewModel : BaseViewModel() {
-}
